DIG-465 Adds sync to table.

+++ docroot/modules/custom/bos_content/modules/node_buildinghousing/src/BuildingHousingUtils.php
@@ -163,6 +163,7 @@ class BuildingHousingUtils {
 
       return $projectWebLink;
     }
+    return "";
   }
 
   /**
@@ -662,7 +663,7 @@ class BuildingHousingUtils {
     //    drush salesforce_pull:pull-query bh_website_update --where="Project__c='a040y00000Z88KlAAJ'" --force-pull
     //    drush salesforce_pull:pull-query building_housing_project_update --where="Project__c='a040y00000Z88KlAAJ'" --force-pull
     //    drush salesforce_pull:pull-query bh_community_meeting_event --where="website_update__c = 'a560y000000WIP2AAO'" --force-pull
-    // then to see whats imported:
+    // then to see what's imported:
     //    drush queue:list
     // then to import queue:
     //    drush queue:run cron_salesforce_pull (optionally --items-limit=X to restrict import)
@@ -675,7 +676,7 @@ class BuildingHousingUtils {
 
     $count = 0;
     foreach ($node_storage->loadMultiple($projects) as $bh_project) {
-      self::deleteProject($bh_project, FALSE, $delete, $log);
+      self::deleteProject($bh_project, $delete, $log);
       $count++;
     }
 
@@ -713,17 +714,26 @@ class BuildingHousingUtils {
     $log && self::log("cleanup", "Processing " . count($projects) . " Project Records. \n");
     foreach ($projects as $bh_project) {
       $lock->acquire(SalesforceSyncSettings::lockname, 30);
-      self::deleteProject($bh_project, $delete, $delete, $log);
+      self::deleteProject($bh_project, $delete, $log);
     }
+    unset($projects);
 
     // Delete any orphaned Updates and linked items.
-    $updates = $node_storage->loadByProperties(["type" => "bh_update"]);
+    $updates = $node_storage->getQuery()
+      ->condition("type", "bh_update")
+      ->execute();
     $log && self::log("cleanup", "\nThere are " . count($updates) . " orphaned Project Updates. \n");
-    foreach ($updates as $bh_update) {
-      $lock->acquire(SalesforceSyncSettings::lockname, 30);
-      self::deleteUpdate(NULL, $bh_update, $delete, $log);
+    if ($delete) {
+      foreach (array_chunk($updates, 500) as $chunk) {
+        $upds = $node_storage->loadMultiple($chunk);
+        foreach ($upds as $bh_update) {
+          $lock->acquire(SalesforceSyncSettings::lockname, 30);
+          self::deleteUpdate(NULL, $bh_update, $delete, $log);
+        }
+      }
+      unset($upds);
     }
-    $updates = NULL;
+    unset($updates);
 
     // Delete orphaned Parcel Assocs.
     $parcel_assocs = $node_storage->getQuery()
@@ -734,29 +744,48 @@ class BuildingHousingUtils {
       foreach (array_chunk($parcel_assocs, 500) as $chunk) {
         $lock->acquire(SalesforceSyncSettings::lockname, 300);
         if (!empty($chunk) && count($chunk) >= 1) {
-          self::deleteParcelAssoc($chunk, $delete, $log, FALSE);
+          self::deleteParcelAssoc($chunk, $delete, $log);
         }
       }
+      unset($chunk);
+      $log && self::log("cleanup", "  DELETED " . count($parcel_assocs) . " parcel-project associations. \n");
     }
-    $log && self::log("cleanup", "  DELETED " . count($parcel_assocs) . " parcel-project associations. \n");
+    unset($parcel_assocs);
 
     // Delete orphaned Parcels.
-    $config = \Drupal::config('node_buildinghousing.settings');
-    if ($config->get('delete_parcel') ?? FALSE) {
-      $parcels = $node_storage->getQuery()
-        ->condition("type", "bh_parcel")
-        ->execute();
-      $log && self::log("cleanup", "\nThere are " . count($parcels) . " orphaned parcels. \n");
-      if ($delete) {
-        foreach (array_chunk($parcels, 500) as $chunk) {
-          $lock->acquire(SalesforceSyncSettings::lockname, 300);
-          if (!empty($chunk) && count($chunk) >= 1) {
-            self::deleteParcel($chunk, $delete, $log);
-          }
+    $parcels = $node_storage->getQuery()
+      ->condition("type", "bh_parcel")
+      ->execute();
+    $log && self::log("cleanup", "\nThere are " . count($parcels) . " orphaned parcels. \n");
+    if ($delete) {
+      foreach (array_chunk($parcels, 500) as $chunk) {
+        $lock->acquire(SalesforceSyncSettings::lockname, 300);
+        if (!empty($chunk) && count($chunk) >= 1) {
+          self::deleteParcel($chunk, $delete, $log);
         }
-        $log && self::log("cleanup", "    DELETED " . count($chunk) . " parcel records.\n");
       }
+      $log && self::log("cleanup", "    DELETED " . count($parcels) . " parcel records.\n");
+      unset($chunk);
     }
+    unset($parcels);
+
+    // Delete orphaned Meetings.
+    $meetings = $node_storage->getQuery()
+      ->condition("type", "bh_meeting")
+      ->execute();
+    $log && self::log("cleanup", "\nThere are " . count($meetings) . " orphaned meetings. \n");
+    if ($delete) {
+      foreach (array_chunk($meetings, 500) as $chunk) {
+        $mtgs = $node_storage->loadMultiple($chunk);
+        foreach($mtgs as $mtg) {
+          $lock->acquire(SalesforceSyncSettings::lockname, 30);
+          self::deleteMeeting($mtg, $delete, $log);
+        }
+      }
+      unset($chunk);
+      $log && self::log("cleanup", "    DELETED " . count($meetings) . " meeeting records.\n");
+    }
+    unset($meetings);
 
     $log && self::log("cleanup", "===CLEANUP ends\n");
 
@@ -776,7 +805,7 @@ class BuildingHousingUtils {
    * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
    * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
    */
-  public static function deleteProject($bh_project, $parcel, $delete, $log) {
+  public static function deleteProject($bh_project, $delete, $log) {
 
     $log && self::log("cleanup", "Scanning PROJECT {$bh_project->getTitle()} ({$bh_project->id()})\n");
 
@@ -808,9 +837,6 @@ class BuildingHousingUtils {
       self::deleteFile($file, [$bh_project->id()], $delete, $log);
     }
 
-    $config = \Drupal::config('node_buildinghousing.settings');
-    $parcel_delete = $delete && ($config->get('delete_parcel') ?? FALSE);
-
     // Find Parcel-Project Associations and delete those.
     if ($bh_project->get('field_bh_parcel_id')->value) {
       foreach ($node_storage->loadByProperties([
@@ -819,20 +845,18 @@ class BuildingHousingUtils {
       ]) as $bh_parcel_assoc) {
         // Note deletes parcel-project association mapping objects, and also
         // the parcel objects referenced by the association mappings.
-        self::deleteParcelAssoc($bh_parcel_assoc, $delete, $log, $parcel_delete);
+        self::deleteParcelAssoc($bh_parcel_assoc, $delete, $log);
       }
     }
 
     // If the bh_project itself is linked to a parcel, then delete that now.
-    if ($parcel_delete) {
-      // Find associated Parcels and delete those.
-      if ($parcel && $bh_project->get('field_bh_parcel_id')->value) {
-        foreach ($node_storage->loadByProperties([
-          "type" => "bh_parcel",
-          "title" => $bh_project->get('field_bh_parcel_id')->value,
-        ]) as $bh_parcel) {
-          self::deleteParcel($bh_parcel, $parcel_delete, $log);
-        }
+    // Find associated Parcels and delete those.
+    if ($bh_project->get('field_bh_parcel_id')->value) {
+      foreach ($node_storage->loadByProperties([
+        "type" => "bh_parcel",
+        "title" => $bh_project->get('field_bh_parcel_id')->value,
+      ]) as $bh_parcel) {
+        self::deleteParcel($bh_parcel, $delete, $log);
       }
     }
 
@@ -935,20 +959,19 @@ class BuildingHousingUtils {
 
   }
 
-  public static function deleteParcelAssoc($bh_parcel_assoc, $delete, $log, $del_parcel) {
+  public static function deleteParcelAssoc($bh_parcel_assoc, $delete, $log) {
 
     if ($delete) {
       if (is_array($bh_parcel_assoc)) {
         $entities = \Drupal::entityTypeManager()->getStorage("node")->loadMultiple($bh_parcel_assoc);
-        $c = count($entities);
         // Delete any parcels which are linked to this assoc.
-        if ($del_parcel) {
-          $parcels = [];
-          foreach ($entities as $entity) {
+        $parcels = [];
+        foreach ($entities as $entity) {
+          if ($entity->hasField("bh_parcel_ref")) {
             $parcels[] = $entity->fields("bh_parcel_ref")->value;
           }
-          self::deleteParcel($parcels, $delete, $log);
         }
+        self::deleteParcel($parcels, $delete, $log);
 
         \Drupal::entityTypeManager()->getStorage("node")->delete($entities);
         unset($entities);
@@ -957,6 +980,9 @@ class BuildingHousingUtils {
         $bh_parcel_assoc->delete();
         $log && self::log("cleanup", "  DELETED PROJECT-PARCEL ASSOC {$bh_parcel_assoc->getTitle()}\n");
       }
+    }
+    else {
+      $log && self::log("cleanup", "    Dry-run {$bh_parcel_assoc->getTitle()} NOT DELETED\n");
     }
 
   }
@@ -969,7 +995,7 @@ class BuildingHousingUtils {
    * @param $file FileEntity The file object we are checking if we can delete.
    * @param $ids array An array of "parent" entity ID's we are going to delete.
    * @param $delete bool Flag. FALSE = dry-run (nothing deleted)
-   * @param $log bool Should we log this deletion.
+   * @param $log bool Should we log this deletion?
    *
    * @return void
    */
@@ -1079,7 +1105,7 @@ class BuildingHousingUtils {
     $body = html_entity_decode(str_replace("&nbsp;", " ", htmlentities($body)));
     $body = str_replace("<br>", " ", $body);
     $body = preg_replace("/\s{2,}/", " ", $body);
-    // Ensure plain-text, in-line URL's are properly wrapped with anchors.
+    // Ensure plain-text, in-line URLs are properly wrapped with anchors.
     $body = preg_replace(['/([^"\'])(http[s]?:.*?)([\s\<]|$)/'], ['${1}<a href="${2}">${2}</a>${3}'], $body);
     return $body;
   }
